services/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["yabindra-bhujel <yabindra2057@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
amqp = "5.2.0"
asgiref = "3.7.2"
astroid = "2.15.8"
async-timeout = "4.0.3"
attrs = "23.1.0"
autobahn = "23.6.2"
automat = "22.10.0"
beautifulsoup4 = "4.12.2"
bidict = "0.22.1"
billiard = "4.2.0"
blinker = "1.7.0"
brotli = "1.1.0"
bs4 = "0.0.1"
cdifflib = "1.2.6"
celery = "5.3.6"
certifi = "2023.7.22"
cffi = "1.16.0"
channels = "4.0.0"
charset-normalizer = "3.2.0"
click = "8.1.7"
click-didyoumean = "0.3.0"
click-plugins = "1.1.1"
click-repl = "0.3.0"
constantly = "15.1.0"
cryptography = "41.0.4"
daphne = "4.0.0"
decorator = "4.4.2"
dill = "0.3.7"
dj-database-url = "2.1.0"
django = "4.2.4"
django-cors-headers = "4.2.0"
django-filter = "23.2"
django-multiupload = "0.6.1"
django-tenant-schemas = "1.11.0"
django-tenants = "3.5.0"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.2.2"
dnspython = "2.4.2"
drf-spectacular = "0.26.5"
drf-yasg = "1.21.7"
eventlet = "0.33.3"
flask = "3.0.0"
flask-socketio = "5.3.6"
fuzzywuzzy = "0.18.0"
greenlet = "3.0.0"
gunicorn = "21.2.0"
h11 = "0.14.0"
hyperlink = "21.0.0"
idna = "3.4"
imageio = "2.31.3"
imageio-ffmpeg = "0.4.9"
incremental = "22.10.0"
inflection = "0.5.1"
iniconfig = "2.0.0"
isort = "5.12.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
jsonschema = "4.20.0"
jsonschema-specifications = "2023.11.2"
kombu = "5.3.5"
lazy-object-proxy = "1.9.0"
markdown = "3.4.4"
markupsafe = "2.1.3"
mccabe = "0.7.0"
moviepy = "1.0.3"
numpy = "1.26.3"
ordered-set = "4.1.0"
packaging = "23.1"
pillow = "9.5.0"
platformdirs = "3.10.0"
pluggy = "1.3.0"
proglog = "0.1.10"
prompt-toolkit = "3.0.43"
psycopg2 = "2.9.7"
psycopg2-binary = "2.9.7"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pyflakes = "3.1.0"
pyjwt = "2.8.0"
pylint = "2.17.7"
pyopenssl = "23.2.0"
pyotp = "2.9.0"
pypng = "0.20220715.0"
pytest = "7.4.2"
pytest-django = "4.5.2"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-engineio = "4.8.1"
pytz = "2023.3"
pyyaml = "6.0.1"
qrcode = "7.4.2"
redis = "5.0.2"
referencing = "0.32.0"
reportlab = "4.0.5"
requests = "2.31.0"
rpds-py = "0.13.2"
service-identity = "23.1.0"
simple-websocket = "1.0.0"
six = "1.16.0"
soupsieve = "2.4.1"
sqlparse = "0.4.4"
subprocess-run = "0.0.8"
tomlkit = "0.12.1"
tqdm = "4.66.1"
twisted = "23.8.0"
txaio = "23.1.1"
typing-extensions = "4.7.1"
tzdata = "2024.1"
uritemplate = "4.1.1"
urllib3 = "2.0.4"
uwsgi = "2.0.23"
vine = "5.1.0"
wcwidth = "0.2.13"
websocket-client = "1.6.3"
websockets = "11.0.3"
werkzeug = "3.0.1"
whitenoise = "6.6.0"
wrapt = "1.15.0"
wsproto = "1.2.0"
zope-interface = "6.1"
python-socketio = "5.10.0"
django-debug-toolbar = "4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
